+++ app/(base)/member/daily-routine/components/DailyMessage.module.scss
@@ -1,0 +1,16 @@
+.wrapper {
+    display: flex;
+    flex-direction: column;
+    padding: 24px;
+    max-width: 900px;
+    width: 100%;
+
+    margin: 0 auto;
+
+    > h2 {
+        font-size: var(--fontSize-2xl);
+        font-weight: var(--fontWeight-bold);
+        text-align: center;
+        margin-bottom: 16px;
+    }
+}
